+++ tests/test_unit_app.py
@@ -1,15 +1,12 @@
-import sys
-
 import pytest
-from nicegui import ui
 from nicegui.testing import User
 
-from tdamapper import _app
+from tdamapper import app
 
 pytest_plugins = ["nicegui.testing.user_plugin"]
 
 
-@pytest.mark.module_under_test(_app)
+@pytest.mark.module_under_test(app)
 async def test_run_mapper(user: User) -> None:
     await user.open("/")
     await user.should_see("Load Data")
